+++ felelj/src/main/java/hu/dorin/felelj/repository/ChoiceRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import hu.dorin.felelj.model.Choice;
 
 @RepositoryRestResource
-public interface ChoiceRepository extends CrudRepository<Choice, Integer> {
+public interface ChoiceRepository extends CrudRepository<Choice, Long> {
 
 }
